backend/src/requirements/base.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
pyotp==2.6.0
python-dotenv==1.0.0
drf-yasg==1.21.5
celery==5.2.7
redis==4.5.3
flower==1.2.0
crontab==1.0.1